refactor: modal to atom component
