+++ resources/js/Components/Atom/index.ts
@@ -3,6 +3,7 @@ export {default as Button} from './Button'
 export {default as DataTable} from './DataTable'
 export {default as Drawer} from './Drawer'
 export {default as Link} from './Link'
+export {default as Modal} from './Modal'
 export {default as NumberInput} from './NumberInput'
 export {default as Select} from './Select'
 export {default as StepperForm} from './StepperForm'
